chore(engine): add workflow package
The new workflow package introduces an abstraction over physical plans:

* A physical plan is split into several parallelizable units called
  "tasks."

* Each task sends or receives data along streams.

* The "workflow" is the graph of tasks to execute for a given query. A
  workflow listens for task results from the root task.

Other engines typically call these units "fragments," with the
collection of fragments constructing the "distributed query plan."

We don't use the standard term here since our usage is stateful.
Workflows respond to tasks changing state, and will eventually be
responsible for deciding when a task should be enqueued for running at
all.

The workflow.Runner interface is introduced to represent the mechanism
running tasks. A basic implementation is used for testing, but in
production, workflow.Runner will be implemented by the scheduler.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>